Release FUT-2.4

+++ pod/generic/artifactory_reader.py
@@ -1,19 +1,25 @@
 import re
+import os
 import requests
 from pathlib import Path
 
+from lib_testbed.generic.util.logger import log
 from lib_testbed.generic.util.common import wait_for
 from lib_testbed.generic.util.artifactory_lib import get_artifactory_fw_url, get_map, get_enc_key
+
+
+DEFAULT_BUILD_MAP = "build_map.json"
 
 
 class ArtifactoryReader:
     downloading = False
 
-    def __init__(self, lib, model=None, artifactory_cfg=None):
+    def __init__(self, lib, model=None, artifactory_cfg=None, map_type: str = DEFAULT_BUILD_MAP):
         self.lib = lib
         self.initialized = False
         self.model = model
         self.build_map = None
+        self.map_type = map_type
         self.artifactory = artifactory_cfg if artifactory_cfg else self.lib.config.get("artifactory")
         self.version_map = {"version": "", "build_num": "", "native": "", "legacy": ""}
         self.header = {"Content-Type": "application/json"}
@@ -25,8 +31,8 @@ class ArtifactoryReader:
         if self.artifactory is None:
             raise Exception("Missing artifactory section in config, cannot move forward")
         if self.model is None:
-            self.model = self.lib.device.config['model_org']
-        self.build_map = get_map(self.model)
+            self.model = self.lib.device.config["model_org"]
+        self.build_map = get_map(self.model, map_type=self.map_type)
         Path(self.tmp_dir).mkdir(exist_ok=True, parents=True)
         self.initialized = True
 
@@ -34,10 +40,10 @@ class ArtifactoryReader:
         self.initialize()
         parsed_data = {}
         if (
-                "master" in requested_version
-                or "fbb" in requested_version
-                or "native" in requested_version
-                or "legacy" in requested_version
+            "master" in requested_version
+            or "fbb" in requested_version
+            or "native" in requested_version
+            or "legacy" in requested_version
         ):
             splitted_version_to_build = requested_version.split("-")
         else:
@@ -61,18 +67,19 @@ class ArtifactoryReader:
 
         self.version_map.update(parsed_data)
 
-    def get_list_of_files(self, version: dict) -> tuple:
+    def get_list_of_files(self, version: dict, final_version: str = None) -> tuple:
         def build_sort_func(build_name):
             build_name_separated = build_name.split("-")
-            version_from_regex = re.search(r"\d+.\d+.\d+", build_name).group(0)
+            version_from_regex = re.search(r"\d+\.\d+\.\d+(\.\d+)?", build_name).group(0)
             version_index = build_name_separated.index(version_from_regex)
 
             return int(build_name_separated[version_index + 1])
 
         self.initialize()
         fw_list = []
-        final_version = version["native"] + "-" + version["version"] if version["native"] else version["version"]
-        final_version = version["legacy"] + "-" + final_version if version["legacy"] else final_version
+        if not final_version:
+            final_version = version["native"] + "-" + version["version"] if version["native"] else version["version"]
+            final_version = version["legacy"] + "-" + final_version if version["legacy"] else final_version
         storage_url = self.artifactory["url"] + f'/api/storage/{self.build_map[final_version]["proj-name"]}'
 
         if self.artifactory.get("user") and self.artifactory.get("password"):
@@ -131,15 +138,29 @@ class ArtifactoryReader:
     def stop_downloading_flag(cls):
         cls.downloading = False
 
-    def download_proper_version(self, version: str, **kwargs) -> str:
+    def download_proper_version(self, version: str, **kwargs) -> str | None:
         destination_dir = kwargs.pop("destination", self.tmp_dir)
         self.initialize()
         url = self.get_url_for_fw(version, **kwargs)
-        filename = self.get_filename_from_url(url)
-        if Path(destination_dir).joinpath(filename).is_file():
+        try:
+            filename = self.get_filename_from_url(url)
+        except (AttributeError, IndexError):
+            log.error("Cannot extract filename from %s", url)
+            return None
+        #  The conditions below must work with the following use cases:
+        #  1. Doing a regular pod upgrade, download one firmware for all 3 pods.
+        #  2. Doing a pod upgrade-multi, download one firmware per version specified by the user.
+        #  3. This method is invoked at roughly the same time, i.e. all downloads must start before at least one
+        #     of them is finished - otherwise this logic might not work as expected. It's because the very first
+        #     finished download sets cls.downloading to False [the very first thread that is done]. This might be the
+        #     case when downloading multiple firmware versions in parallel.
+        if Path(destination_dir).joinpath(filename).is_file() and self.is_not_downloading():
             return filename
-        if not self.is_downloading():
+        if not (os.path.exists(f"{destination_dir}/{filename}") and self.is_downloading()):
             self.start_downloading_flag()
+            with open(f"{destination_dir}/{filename}", "wb") as file:
+                file.write(b"")  # make sure that empty file exists
+            log.info(f"Downloading {url} to {self.tmp_dir} directory")
             downloaded_image = requests.get(url)
             with open(f"{destination_dir}/{filename}", "wb") as file:
                 file.write(downloaded_image.content)
@@ -157,7 +178,7 @@ class ArtifactoryReader:
             self.version_map["version"] = "build_device_featurebranch"
             version = f"{legacy}{native}{self.version_map['version']}-{self.version_map['build_num']}"
         if self.version_map["version"] and self.version_map["build_num"]:
-            url = get_artifactory_fw_url(self.lib.config, version, self.model, **kwargs)
+            url = get_artifactory_fw_url(self.lib.config, version, self.model, self.map_type, **kwargs)
         else:
             url = (
                 f'{self.artifactory["url"]}/'
